graph: graph builder now uses workspaces
