+++ graph/graph.go
@@ -6,15 +6,13 @@
 package graph
 
 import (
-	"fmt"
 	"log"
 	"os"
-
-	"sync"
 
 	"bldy.build/build"
 	"bldy.build/build/label"
 	"bldy.build/build/skylark"
+	"bldy.build/build/workspace"
 	"github.com/pkg/errors"
 
 	"bldy.build/build/postprocessor"
@@ -25,41 +23,18 @@ var (
 	l = log.New(os.Stdout, "graph: ", 0)
 )
 
-// Node encapsulates a target and represents a node in the build graph.
-type Node struct {
-	IsRoot        bool       `json:"-"`
-	Target        build.Rule `json:"-"`
-	Type          string
-	Parents       map[string]*Node `json:"-"`
-	Label         label.Label
-	Worker        string
-	PriorityCount int
-	WG            sync.WaitGroup
-	Status        build.Status
-	Cached        bool
-	Start, End    int64
-	Hash          string
-	Output        string `json:"-"`
-	Once          sync.Once
-	sync.Mutex
-	Children map[string]*Node
-	hash     []byte
-}
-
-// Graph represents a build graph
-type Graph struct {
-	Root  *Node
-	vm    build.VM
-	Nodes map[string]*Node
-}
-
 // New returns a new build graph relatvie to the working directory
 func New(wd, target string) (*Graph, error) {
-	vm, err := skylark.New(wd)
+	ws, err := workspace.New(wd)
 	if err != nil {
-		return nil, errors.Wrap(err, "new graph")
+		return nil, errors.Wrap(err, "graph: new")
+	}
+	vm, err := skylark.New(ws)
+	if err != nil {
+		return nil, errors.Wrap(err, "graph: new")
 	}
 	g := Graph{
+		ws:    ws,
 		vm:    vm,
 		Nodes: make(map[string]*Node),
 	}
@@ -72,45 +47,37 @@ func New(wd, target string) (*Graph, error) {
 	return &g, nil
 }
 
-// Priority counts how many nodes directly and indirectly depend on
-// this node
-func (n *Node) Priority() int {
-	if n.PriorityCount < 0 {
-		p := 0
-		for _, c := range n.Parents {
-			p += c.Priority() + 1
-		}
-		n.PriorityCount = p
-	}
-	return n.PriorityCount
+// Graph represents a build graph
+type Graph struct {
+	Root  *Node
+	vm    build.VM
+	ws    workspace.Workspace
+	Nodes map[string]*Node
+}
+
+// Workspace returns the Workspace in which this graph exists.
+func (g *Graph) Workspace() workspace.Workspace {
+	return g.ws
 }
 
 func (g *Graph) getTarget(lbl *label.Label) (n *Node) {
 	if gnode, ok := g.Nodes[lbl.String()]; ok {
 		return gnode
 	}
+
 	t, err := g.vm.GetTarget(lbl)
 	if err != nil {
 		l.Fatal(err)
 	}
+
 	nLbl := label.Label{
 		Package: lbl.Package,
 		Name:    t.GetName(),
 	}
 
-	node := Node{
-		Target:        t,
-		Type:          fmt.Sprintf("%T", t)[1:],
-		Children:      make(map[string]*Node),
-		Parents:       make(map[string]*Node),
-		Once:          sync.Once{},
-		WG:            sync.WaitGroup{},
-		Status:        build.Pending,
-		Label:         nLbl,
-		PriorityCount: -1,
-	}
+	node := NewNode(nLbl, t)
 
-	post := postprocessor.New(nLbl.Package)
+	post := postprocessor.New(g.ws, nLbl)
 
 	err = post.ProcessDependencies(node.Target)
 	if err != nil {
@@ -126,11 +93,12 @@ func (g *Graph) getTarget(lbl *label.Label) (n *Node) {
 		group = node.Target.(*bldytrg.Group)
 		group.Exports = make(map[string]string)
 	}
+
 	for _, d := range node.Target.GetDependencies() {
 		depLbl, err := label.Parse(d)
 		c := g.getTarget(depLbl)
 		if err != nil {
-			l.Printf("%q is not a valid url", depLbl)
+			l.Printf("%q is not a valid label", depLbl)
 			continue
 		}
 		node.WG.Add(1)
