+++ graph/graph_test.go
@@ -1,0 +1,7 @@
+package graph
+
+import "testing"
+
+func TestNew(t *testing.T) {
+
+}
